requirements.txt
Django==4.0.1
djangorestframework==3.13.1
djangorestframework_simplejwt==5.0.0
psycopg2-binary
django-filter